[Updating] Use PowerToys.Update for update procedures
