+++ src/common/updating/installer.h
@@ -10,7 +10,7 @@
 namespace updating
 {
     std::wstring get_msi_package_path();
-    bool uninstall_msi_version(const std::wstring& package_path, const notifications::strings&);
+    bool uninstall_msi_version(const std::wstring& package_path);
     bool offer_msi_uninstallation(const notifications::strings&);
     std::optional<std::wstring> get_msi_package_installed_path();
