+++ src/runner/ActionRunnerUtils.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#define WIN32_LEAN_AND_MEAN
+#include <Windows.h>
+
+namespace cmdArg
+{
+    const inline wchar_t* RUN_NONELEVATED = L"-run-non-elevated";
+}
