+++ src/runner/UpdateUtils.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#define WIN32_LEAN_AND_MEAN
-#include <Windows.h>
+#include <common/updating/updating.h>
 
-SHELLEXECUTEINFOW launch_action_runner(const wchar_t* cmdline);
+void PeriodicUpdateWorker();
+void CheckForUpdatesSettingsCallback();
 
 namespace cmdArg
 {
@@ -16,7 +16,7 @@ namespace cmdArg
     const inline wchar_t* UPDATE_STAGE2_RESTART_PT = L"restart";
     const inline wchar_t* UPDATE_STAGE2_DONT_START_PT = L"dont_start";
 
-    const inline wchar_t* RUN_NONELEVATED = L"-run-non-elevated";
-
     const inline wchar_t* UPDATE_REPORT_SUCCESS = L"-report_update_success";
 }
+
+SHELLEXECUTEINFOW LaunchPowerToysUpdate(const wchar_t* cmdline);
