+++ src/runner/runner.vcxproj
@@ -46,7 +46,6 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\common\interop\two_way_pipe_message_ipc.cpp" />
-    <ClCompile Include="action_runner_utils.cpp" />
     <ClCompile Include="auto_start_helper.cpp" />
     <ClCompile Include="CentralizedHotkeys.cpp" />
     <ClCompile Include="general_settings.cpp" />
@@ -62,17 +61,17 @@
     <ClCompile Include="trace.cpp" />
     <ClCompile Include="tray_icon.cpp" />
     <ClCompile Include="unhandled_exception_handler.cpp" />
-    <ClCompile Include="update_utils.cpp" />
+    <ClCompile Include="UpdateUtils.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="action_runner_utils.h" />
+    <ClInclude Include="ActionRunnerUtils.h" />
     <ClInclude Include="auto_start_helper.h" />
     <ClInclude Include="CentralizedHotkeys.h" />
     <ClInclude Include="general_settings.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="centralized_kb_hook.h" />
     <ClInclude Include="settings_telemetry.h" />
-    <ClInclude Include="update_utils.h" />
+    <ClInclude Include="UpdateUtils.h" />
     <ClInclude Include="powertoy_module.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="restart_elevated.h" />
