--- src/runner/update_utils.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <common/updating/updating.h>
-
-bool start_msi_uninstallation_sequence();
-void periodic_update_worker();
-void check_for_updates_settings_callback();
